+++ app/controllers/application_controller.rb
@@ -7,10 +7,11 @@ class ApplicationController < ActionController::Base
   private
 
   def logged_in?
-    !!session[:user_id]
+    !!current_user
   end
 
   def current_user
+    return unless session[:user_id]
     @current_user ||= User.find_by(id: session[:user_id]) if session[:user_id]
   end
