+++ app/controllers/subscriptions_controller.rb
@@ -5,6 +5,7 @@ class SubscriptionsController < ApplicationController
 
   def index
     @subscriptions = current_user.subscriptions
+    render 
   end
 
   def show; end
