+++ app/controllers/users_controller.rb
@@ -1,0 +1,5 @@
+class UsersController < ApplicationController
+  def index
+    @user = current_user
+  end
+end
