+++ app/views/subscriptions/index.html.slim
@@ -4,7 +4,7 @@
         = button_to '新規登録', new_subscription_path, method: :get
   .flex.items-center.flex-shrink-0
     .bg-blue-500.text-sm.md:text-base.hover:bg-blue-700.text-white.font-bold.py-2.px-4.rounded
-        = button_to 'カレンダー連携', subscriptions_calendars_url(protocol: :webcal, format: :ics)
+        = button_to 'カレンダー連携', user_calendars_url(protocol: :webcal, format: :ics)
 
 - if @subscriptions.blank?
   a.block.max-w.my-5.p-6.bg-white.border.border-gray-200.rounded-lg.shadow.hover:bg-gray-100.dark:bg-gray-800.dark:border-gray-700.dark:hover:bg-gray-700
