+++ config/application.rb
@@ -23,6 +23,7 @@ module SubscMine
     # Initialize configuration defaults for originally generated Rails version.
     config.load_defaults 7.0
     config.i18n.default_locale = :ja
+    config.middleware.use ActionDispatch::Cookies
     # Configuration for the application, engines, and railties goes here.
     #
     # These settings can be overridden in specific environments using the files
