+++ config/routes.rb
@@ -6,14 +6,18 @@ Rails.application.routes.draw do
   get 'signout', to: 'sessions#destroy', as: 'signout'
 
   resource :retirements
-
+  
+  resources :users, only: %i(index)
   resources :sessions, only: %i(new create destroy)
   resources :welcome, only: %i(index)
   resources :subscriptions, only: %i(index new create edit update destroy)
 
   namespace :subscriptions do
-    resources :calendars, only: %i(index)
     resources :status, only: %i(edit update)
+  end
+  
+  resources :users do
+    resources :calendars, only: %i(index)
   end
   # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html
